Move check_polls to MeetingUserMergeMixin
