+++ openslides_backend/action/actions/user/merge_mixins.py
@@ -13,6 +13,7 @@ from ....models.models import (
     Poll,
     Speaker,
 )
+from ....services.database.commands import GetManyRequest
 from ....shared.exceptions import ActionException
 from ....shared.filters import And, FilterOperator, Or
 from ....shared.patterns import Collection, fqid_from_collection_and_id
@@ -279,11 +280,37 @@ class MeetingUserMergeMixin(
             case _:
                 return super().get_merge_comparison_hash(collection, model)
 
-    def check_polls_helper(self, meeting_users: dict[int, dict[str, Any]]) -> list[str]:
+    def check_polls(self, into: PartialModel, other_models: list[PartialModel]) -> None:
         messages: list[str] = []
         group_ids: set[int] = set()
         meeting_ids: set[int] = set()
         meeting_id_by_group_ids: dict[int, int] = {}
+
+        meeting_user_ids = [
+            mu_id
+            for user in [into, *other_models]
+            for mu_id in user.get("meeting_user_ids", [])
+        ]
+        meeting_users = self.datastore.get_many(
+            [
+                GetManyRequest(
+                    "meeting_user",
+                    meeting_user_ids,
+                    [
+                        "user_id",
+                        "meeting_id",
+                        "group_ids",
+                        "vote_delegations_from_ids",
+                        "vote_delegated_to_id",
+                        "poll_voted_ids",
+                        "poll_option_ids",
+                        "acting_ballot_ids",
+                        "represented_ballot_ids",
+                    ],
+                )
+            ]
+        )["meeting_user"]
+
         for m_user in meeting_users.values():
             if len(g_ids := m_user.get("group_ids", [])):
                 meeting_id_by_group_ids.update(
@@ -360,4 +387,71 @@ class MeetingUserMergeMixin(
             messages.append(
                 f"some of the selected users are delegating votes to each other in meeting(s) {', '.join(bad_meetings)}"
             )
-        return messages
+
+        ballot_poll_ids_per_user_id: dict[int, set[int]] = {}
+        option_poll_ids_per_user_id: dict[int, set[int]] = {}
+        for meeting_user in meeting_users.values():
+            if poll_voted_ids := meeting_user.get("poll_voted_ids"):
+                ballot_poll_ids_per_user_id.setdefault(
+                    meeting_user["user_id"], set()
+                ).update(set(poll_voted_ids))
+        for meeting_user in meeting_users.values():
+            if len(
+                (o_ids := meeting_user.get("poll_option_ids", []))
+                + (
+                    b_ids := list(
+                        {
+                            id_
+                            for id_ in [
+                                *meeting_user.get("acting_ballot_ids", []),
+                                *meeting_user.get("represented_ballot_ids", []),
+                            ]
+                        }
+                    )
+                )
+            ):
+                get_many_requests = [
+                    GetManyRequest("poll_option", o_ids, ["poll_id"]),
+                    GetManyRequest("poll_ballot", b_ids, ["poll_id"]),
+                ]
+                many_models = self.datastore.get_many(get_many_requests)
+                if o_ids:
+                    option_poll_ids_per_user_id.setdefault(
+                        meeting_user["user_id"], set()
+                    ).update(
+                        {
+                            option["poll_id"]
+                            for option in many_models["poll_option"].values()
+                        }
+                    )
+                ballot_data = many_models["poll_ballot"]
+                ballot_poll_ids_per_user_id.setdefault(
+                    meeting_user["user_id"], set()
+                ).update({ballot["poll_id"] for ballot in ballot_data.values()})
+        ballot_conflicts = {
+            poll_id
+            for id1, poll_ids1 in ballot_poll_ids_per_user_id.items()
+            for id2, poll_ids2 in ballot_poll_ids_per_user_id.items()
+            for poll_id in poll_ids1.intersection(poll_ids2)
+            if id1 != id2
+        }
+        option_conflicts = {
+            poll_id
+            for id1, poll_ids1 in option_poll_ids_per_user_id.items()
+            for id2, poll_ids2 in option_poll_ids_per_user_id.items()
+            for poll_id in poll_ids1.intersection(poll_ids2)
+            if id1 != id2
+        }
+        if len(ballot_conflicts):
+            messages.append(
+                f"among the selected users multiple voted in poll(s) {', '.join([str(id_) for id_ in ballot_conflicts])}"
+            )
+        if len(option_conflicts):
+            messages.append(
+                f"multiple of the selected users are among the options in poll(s) {', '.join([str(id_) for id_ in option_conflicts])}"
+            )
+
+        if len(messages):
+            raise ActionException(
+                f"Cannot carry out merge into user/{into['id']}, because {' and '.join(messages)}"
+            )
