+++ openslides_backend/action/actions/user/merge_together.py
@@ -359,101 +359,6 @@ class UserMergeTogether(
                 [{"id": id_} for id_ in to_delete],
             )
 
-    def check_polls(self, into: PartialModel, other_models: list[PartialModel]) -> None:
-        all_models = [into, *other_models]
-        meeting_user_ids = [
-            mu_id for user in all_models for mu_id in user.get("meeting_user_ids", [])
-        ]
-        poll_conflict_fields = [
-            "poll_voted_ids",
-            "poll_option_ids",
-            "acting_ballot_ids",
-            "represented_ballot_ids",
-        ]
-        helper_fields = [
-            "vote_delegations_from_ids",
-            "vote_delegated_to_id",
-            "meeting_id",
-            "group_ids",
-        ]
-        meeting_users = self.datastore.get_many(
-            [
-                GetManyRequest(
-                    "meeting_user",
-                    meeting_user_ids,
-                    ["user_id"] + poll_conflict_fields + helper_fields,
-                )
-            ]
-        )["meeting_user"]
-
-        ballot_poll_ids_per_user_id: dict[int, set[int]] = {}
-        option_poll_ids_per_user_id: dict[int, set[int]] = {}
-        for meeting_user in meeting_users.values():
-            if poll_voted_ids := meeting_user.get("poll_voted_ids"):
-                ballot_poll_ids_per_user_id.setdefault(
-                    meeting_user["user_id"], set()
-                ).update(set(poll_voted_ids))
-        for meeting_user in meeting_users.values():
-            if len(
-                (o_ids := meeting_user.get("poll_option_ids", []))
-                + (
-                    b_ids := list(
-                        {
-                            id_
-                            for id_ in [
-                                *meeting_user.get("acting_ballot_ids", []),
-                                *meeting_user.get("represented_ballot_ids", []),
-                            ]
-                        }
-                    )
-                )
-            ):
-                get_many_requests = [
-                    GetManyRequest("poll_option", o_ids, ["poll_id"]),
-                    GetManyRequest("poll_ballot", b_ids, ["poll_id"]),
-                ]
-                many_models = self.datastore.get_many(get_many_requests)
-                if o_ids:
-                    option_poll_ids_per_user_id.setdefault(
-                        meeting_user["user_id"], set()
-                    ).update(
-                        {
-                            option["poll_id"]
-                            for option in many_models["poll_option"].values()
-                        }
-                    )
-                ballot_data = many_models["poll_ballot"]
-                ballot_poll_ids_per_user_id.setdefault(
-                    meeting_user["user_id"], set()
-                ).update({ballot["poll_id"] for ballot in ballot_data.values()})
-        ballot_conflicts = {
-            poll_id
-            for id1, poll_ids1 in ballot_poll_ids_per_user_id.items()
-            for id2, poll_ids2 in ballot_poll_ids_per_user_id.items()
-            for poll_id in poll_ids1.intersection(poll_ids2)
-            if id1 != id2
-        }
-        option_conflicts = {
-            poll_id
-            for id1, poll_ids1 in option_poll_ids_per_user_id.items()
-            for id2, poll_ids2 in option_poll_ids_per_user_id.items()
-            for poll_id in poll_ids1.intersection(poll_ids2)
-            if id1 != id2
-        }
-        messages: list[str] = self.check_polls_helper(meeting_users)
-        if len(ballot_conflicts):
-            messages.append(
-                f"among the selected users multiple voted in poll(s) {', '.join([str(id_) for id_ in ballot_conflicts])}"
-            )
-        if len(option_conflicts):
-            messages.append(
-                f"multiple of the selected users are among the options in poll(s) {', '.join([str(id_) for id_ in option_conflicts])}"
-            )
-        if len(messages):
-            raise ActionException(
-                f"Cannot carry out merge into user/{into['id']}, because {' and '.join(messages)}"
-            )
-
     def get_merge_comparison_hash(
         self, collection: Collection, model: PartialModel
     ) -> int | str | tuple[int | str, ...]:
